+++ lib/tests/aes.c
@@ -87,9 +87,9 @@ struct mac_test_data {
 
 
 static int
-encryption(uint8_t *key, uint8_t counter, uint8_t *plaintext, size_t plaintext_len, uint8_t *enc, size_t enc_len) {
+encryption(uint8_t *key, uint8_t counter, uint8_t *plaintext, uint32_t plaintext_len, uint8_t *enc, size_t enc_len) {
   uint8_t e[255] = {0};
-  size_t e_len = sizeof(e);
+  uint32_t e_len = sizeof(e);
   ykpiv_rc rc = scp11_encrypt_data(key, counter, plaintext, plaintext_len, e, &e_len);
 
   ck_assert(rc == YKPIV_OK);
@@ -98,9 +98,9 @@ encryption(uint8_t *key, uint8_t counter, uint8_t *plaintext, size_t plaintext_l
   return EXIT_SUCCESS;
 }
 
-static int decryption(uint8_t *key, uint8_t counter, uint8_t *enc, size_t enc_len, uint8_t *dec, size_t dec_len) {
+static int decryption(uint8_t *key, uint8_t counter, uint8_t *enc, uint32_t enc_len, uint8_t *dec, size_t dec_len) {
   uint8_t d[255] = {0};
-  size_t d_len = sizeof(d);
+  uint32_t d_len = sizeof(d);
   ykpiv_rc rc = scp11_decrypt_data(key, counter, enc, enc_len, d, &d_len);
 
   ck_assert(rc == YKPIV_OK);
@@ -109,7 +109,7 @@ static int decryption(uint8_t *key, uint8_t counter, uint8_t *enc, size_t enc_le
   return EXIT_SUCCESS;
 }
 
-static int mac(uint8_t *mac_key, uint8_t *mac_chain, uint8_t *data, size_t data_len, uint8_t *mac) {
+static int mac(uint8_t *mac_key, uint8_t *mac_chain, uint8_t *data, uint32_t data_len, uint8_t *mac) {
   uint8_t m[255] = {0};
   ykpiv_rc rc = scp11_mac_data(mac_key, mac_chain, data, data_len, m);
